requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
git+https://github.com/huggingface/diffusers
accelerate>=0.16.0
torchvision
transformers>=4.25.1
ftfy
tensorboard
Jinja2
wandb
bitsandbytes>=0.40.0
xformers>=0.0.20
scipy
fire
gradio